requirements.txt
openai
python-dotenv
python-telegram-bot==20.7
